Create request.rpgle
This is the program to call modules based on the form type defined in the webreq.html used in main.rpgle. As HTML5 supports only GET and POST methods, there is a hidden field _method inside webreq.html forms for PUT and DELETE forms. This _method is then being checked in GET_METHOD procedure in processreq.rpgle module.