Create processreq.rpgle
This module consists of four sub-procedures to process different type of requests received from client. For PUT and DELETE methods, the default method received from form is GET as HTML5 form supports only GET and POST. There is hidden variable _method to simulate PUT and DELETE methods. GET_METHOD will check the value of _method and process the request according to _method value if it is PUT or DELETE. In case of no value received in _method default processing of GET method will happen.